requirements.txt
fastapi
uvicorn
requests
chromadb
python-dotenv
pypdf